+++ src/views/ManageUsers/UserForm.jsx
@@ -3,7 +3,7 @@ import {
   Form,
   Field,
   reduxForm,
-  FieldArray
+  FieldArray, formValueSelector
 } from 'redux-form'
 import { Grid, InputAdornment } from "material-ui";
 import {
@@ -53,15 +53,9 @@ import Stepper from 'material-ui/Stepper';
 import Step from 'material-ui/Stepper/Step'
 import StepLabel from 'material-ui/Stepper/StepLabel';
 import StepContent from 'material-ui/Stepper/StepContent';
-import Paper from 'material-ui/Paper';
-import Typography from 'material-ui/Typography';
-
-const required = value => (value ? undefined : 'Required');
-const isEmail = value => value && !/^[A-Z0-9._%+-]+@devsinc.com$/i.test(value)
-  ? 'Invalid email address, domain must be devsinc.com'
-  : undefined;
-
-
+import UserFormStepOne from './UserFormStepOne';
+import UserFormStepTwo from './UserFormStepTwo';
+import UserFormStepThree from './UserFormStepThree';
 
 function getSteps() {
   return ['Basic Information', 'Employee Record', 'Schedule Email'];
@@ -89,12 +83,6 @@ class UserForm extends React.Component {
     });
   };
   
-  handleReset = () => {
-    this.setState({
-      activeStep: 0,
-    });
-  };
-  
   componentDidMount() {
     const { fetchDepartments, fetchUsers, fetchRoles } = this.props;
     fetchDepartments();
@@ -102,355 +90,66 @@ class UserForm extends React.Component {
     fetchRoles();
   }
   
+  
   render() {
-    const renderFields = ({fields, meta: {error, submitFailed}}) => (
-      <div>
-        Employment History
-        <IconButton
-          color="inherit"
-          aria-label="Add"
-        >
-          <Add
-            onClick={() => fields.push({})}
-          />
-          <Hidden mdUp>
-            <p >Add Position</p>
-          </Hidden>
-        </IconButton>
-        {fields.map((position, index) => (
-          <Grid container key={index}>
-            <ItemGrid xs={4} sm={4} md={4}>
-              <Field name={`${position}.role`} type="text" component={({input,label,...custom}) =>
-                <CustomInput
-                  labelText="Role"
-                  id="role"
-                  formControlProps={{
-                    fullWidth: true
-                  }}
-                  inputProps={{
-                    ...input,
-                    ...custom,
-                    type: "text",
-                    required: "text",
-                    name: `${position}.role`,
-                    autoComplete: "role",
-                  }}
-                />
-              }
-              />
-            </ItemGrid>
-            <ItemGrid xs={3} sm={3} md={3}>
-              <Field name={`${position}.from`} type="number" component={({input,label,...custom}) =>
-                <CustomInput
-                  labelText="From"
-                  id="from"
-                  formControlProps={{
-                    fullWidth: true
-                  }}
-                  inputProps={{
-                    ...input,
-                    ...custom,
-                    type: "text",
-                    required: "text",
-                    name: `${position}.from`,
-                    autoComplete: "from",
-                  }}
-                />
-              }
-              />
-            </ItemGrid>
-            <ItemGrid xs={3} sm={3} md={3}>
-              <Field name={`${position}.to`} type="number" component={({input,label,...custom}) =>
-                <CustomInput
-                  labelText="To"
-                  id="to"
-                  formControlProps={{
-                    fullWidth: true
-                  }}
-                  inputProps={{
-                    ...input,
-                    ...custom,
-                    type: "text",
-                    required: "text",
-                    name: `${position}.to`,
-                    autoComplete: "to",
-                  }}
-                />
-              }
-              />
-            </ItemGrid>
-            <ItemGrid xs={2} sm={2} md={2}>
-              <IconButton
-                color="inherit"
-                aria-label="Remove"
-              >
-                <Remove
-                  onClick={() => fields.remove(index)}
-                />
-              </IconButton>
-            </ItemGrid>
-          </Grid>
-        ))}
-      </div>
-    );
     
     const steps = getSteps();
     const { activeStep } = this.state;
     const { error, handleSubmit, submitting, isNew, users, roles } = this.props;
     return (
-      <Form onSubmit={handleSubmit}>
-        <div>
-          <Stepper activeStep={activeStep} orientation="vertical">
-            {steps.map((label, index) => {
-              return (
-                <Step key={label}>
-                  <StepLabel>{label}</StepLabel>
-                  <StepContent>
-                    {index === 0 &&
-                    <Grid container>
-                      <ItemGrid xs={12} sm={12} md={12}>
-                        <Grid container>
-                          <ItemGrid xs={4} sm={4} md={4}>
-                            <Field validate={[required, isEmail]} name="email" required="required" autoComplete="email" type="email" custominputprops={{labelText: 'Company E-mail'}} component={CustomInputWrapper} />
-                          </ItemGrid>
-                          <ItemGrid xs={4} sm={4} md={4}>
-                            <Field validate={[required]} name="first_name" required="required" autoComplete="first_name" type="text" custominputprops={{labelText: 'First Name'}} component={CustomInputWrapper} />
-                          </ItemGrid>
-                          <ItemGrid xs={4} sm={4} md={4}>
-                            <Field validate={[required]} name="last_name" required="required" autoComplete="last_name" type="text" custominputprops={{labelText: 'Last Name'}} component={CustomInputWrapper} />
-                          </ItemGrid>
-                        </Grid>
-                        <Grid container>
-                          <ItemGrid xs={4} sm={4} md={4}>
-                            <Field name="role_id"
-                                   validate={[required]}
-                                   component={({input}) => (
-                                     <CustomInput
-                                       isSelect={true}
-                                       formControlProps={{
-                                         fullWidth: true
-                                       }}
-                                       labelText={"Role *"}
-                                       inputProps={{
-                                         value: input.value,
-                                         onChange: (event) => {if(event.target.value) return input.onChange(event, event.target.value);},
-                                         required: "required",
-                                         name: "role_id",
-                                         autoComplete: "role_id",
-                                       }}
-                                     >
-                                       {
-                                         roles ?
-                                           roles.map((role, index) => (
-                                               <MenuItem
-                                                 key={index}
-                                                 value={role.id}
-                                               >
-                                                 <ListItemText primary={role.title}/>
-                                               </MenuItem>
-                                             )
-                                           )
-                                           : null
-                                       }
-                                     </CustomInput>
-                                   )
-                                   }
-                            />
-                          </ItemGrid>
-                          <ItemGrid xs={4} sm={4} md={4}>
-                            <Field name="manager_id"
-                                   validate={[required]}
-                                   component={({input}) => (
-                                     <CustomInput
-                                       isSelect={true}
-                                       formControlProps={{
-                                         fullWidth: true
-                                       }}
-                                       labelText={"Manager *"}
-                                       inputProps={{
-                                         value: input.value,
-                                         onChange: (event) => {if(event.target.value) return input.onChange(event, event.target.value);},
-                                         required: "required",
-                                         name: "manager_id",
-                                         autoComplete: "manager_id",
-                                       }}
-                                     >
-                                       {
-                                         users ?
-                                           users.map((user, index) => (
-                                               <MenuItem
-                                                 key={index}
-                                                 value={user.id}
-                                               >
-                                                 <ListItemText primary={user.name}/>
-                                               </MenuItem>
-                                             )
-                                           )
-                                           : null
-                                       }
-                                     </CustomInput>
-                                   )
-                                   }
-                            />
-                          </ItemGrid>
-                          <ItemGrid xs={4} sm={4} md={4}>
-                            <Field name="buddy_id"
-                                   validate={[required]}
-                                   component={({input}) => (
-                                     <CustomInput
-                                       isSelect={true}
-                                       formControlProps={{
-                                         fullWidth: true
-                                       }}
-                                       labelText={"Buddy *"}
-                                       inputProps={{
-                                         value: input.value,
-                                         onChange: (event) => {if(event.target.value) return input.onChange(event, event.target.value);},
-                                         required: "required",
-                                         name: "buddy_id",
-                                         autoComplete: "buddy_id",
-                                       }}
-                                     >
-                                       {
-                                         users ?
-                                           users.map((user, index) => (
-                                               <MenuItem
-                                                 key={index}
-                                                 value={user.id}
-                                               >
-                                                 <ListItemText primary={user.name}/>
-                                               </MenuItem>
-                                             )
-                                           )
-                                           : null
-                                       }
-                                     </CustomInput>
-                                   )
-                                   }
-                            />
-                          </ItemGrid>
-                        </Grid>
-                        <Grid container>
-                          <ItemGrid xs={4} sm={4} md={4}>
-                            <br />
-                            <Field name="joining_date" validate={[required]}  component={(input,label,custom) => (
-                              <DatePicker
-                                label="Joining Date"
-                                {...input}
-                                {...custom}
-                                format="Do MMMM YYYY"
-                                value={input.input.value ? moment(input.input.value) : null }
-                                onChange={(event) => input.input.onChange(event.format("YYYY-MM-DD"))}
-                                disablePast={true}
-                                leftArrowIcon={<KeyboardArrowLeft/>}
-                                rightArrowIcon={<KeyboardArrowRight/>}
-                                InputProps={{
-                                  endAdornment: (
-                                    <InputAdornment position="end">
-                                      <IconButton>
-                                        <DateRange />
-                                      </IconButton>
-                                    </InputAdornment>
-                                  ),
-                                }}
-                              />
-                            )} />
-                          </ItemGrid>
-                        </Grid>
-                      </ItemGrid>
-                    </Grid>
-                    }
-                    
-                    {index === 1 &&
-                    <Grid container>
-                      <ItemGrid xs={12} sm={12} md={12}>
-                        <Grid container>
-                          <ItemGrid xs={4} sm={4} md={4}>
-                            <Field name="address" required="required" autoComplete="address" type="text" custominputprops={{labelText: 'Address'}} component={CustomInputWrapper} />
-                          </ItemGrid>
-                          <ItemGrid xs={4} sm={4} md={4}>
-                            <Field name="emergency_contact" required="required" autoComplete="emergency_contact" type="number" custominputprops={{labelText: 'Emergency Contact'}} component={CustomInputWrapper} />
-                          </ItemGrid>
-                        </Grid>
-                        {
-                          isNew ?
-                            null:
-                            <div>
-                              <h3> Employment History </h3>
-                              <FieldArray name="employment_history" component={renderFields} />
-                            </div>
-                        }
-                      </ItemGrid>
-                    </Grid>
-                    }
-                    
-                    {index === 2 &&
-                    <Grid container>
-                      <ItemGrid xs={4} sm={4} md={4}>
-                        <br />
-                        <Field name="schedule_date_time" validate={[required]} component={(input,label,custom) => (
-                          <DateTimePicker
-                            label="Schedule Welcome E-mail"
-                            {...input}
-                            {...custom}
-                            value={input.input.value ? moment(input.input.value) : null }
-                            disablePast={true}
-                            onChange={(event) => input.input.onChange(event.format("YYYY-MM-DD"))}
-                            timeIcon={<AccessTime/>}
-                            dateRangeIcon={<DateRange/>}
-                            leftArrowIcon={<KeyboardArrowLeft/>}
-                            rightArrowIcon={<KeyboardArrowRight/>}
-                          />
-                        )}
-                        />
-                      </ItemGrid>
-                    </Grid>
-                    }
+      <div>
+        <Stepper activeStep={activeStep} orientation="vertical">
+          {steps.map((label, index) => {
+            return (
+              <Step key={label}>
+                <StepLabel>{label}</StepLabel>
+                <StepContent>
+                  {index === 0 && <UserFormStepOne roles={roles} users={users} onSubmit={handleSubmit} />}
+                  {index === 1 && <UserFormStepTwo onSubmit={handleSubmit}/>}
+                  {index === 2 && <UserFormStepThree onSubmit={handleSubmit}/>}
+                  <div>
                     <div>
-                      <div>
-                        <br />
-                        <br />
+                      <br />
+                      <br />
+                      {
+                        index !== 0 && <Button
+                          disabled={activeStep === 0}
+                          onClick={this.handleBack}
+                        >
+                          Back
+                        </Button>
+                      }
+                      
+                      {
+                        activeStep !== steps.length - 1 &&
+                        <Button
+                          variant="raised"
+                          color="primary"
+                          onClick={() => {this.handleNext()}}
+                        >
+                          {activeStep === steps.length - 1 ? 'Finish' : 'Next'}
+                        </Button>
+                      }
+                      {
+                        activeStep === steps.length - 1 &&
+                        <Button disabled={submitting} onClick={handleSubmit} color="primary">
+                          {isNew ? "Create User" : "Save Changes"}
+                        </Button>
                         
-                        {
-                          index !== 0 && <Button
-                            disabled={activeStep === 0}
-                            onClick={this.handleBack}
-                          >
-                            Back
-                          </Button>
-                        }
-                        
-                        {
-                          activeStep !== steps.length - 1 &&
-                          <Button
-                            variant="raised"
-                            color="primary"
-                            onClick={() => {alert(1); this.handleNext()}}
-                          >
-                            {activeStep === steps.length - 1 ? 'Finish' : 'Next'}
-                          </Button>
-                        }
-                        {
-                          activeStep === steps.length - 1 &&
-                          <Button disabled={submitting} onClick={handleSubmit} color="primary">
-                            {isNew ? "Create User" : "Save Changes"}
-                          </Button>
-                          
-                        }
-                      </div>
+                      }
                     </div>
-                  </StepContent>
-                </Step>
-              );
-            })}
-          </Stepper>
-        </div>
+                  </div>
+                </StepContent>
+              </Step>
+            );
+          })}
+        </Stepper>
         <br/>
         {error
           ? <Danger>{error}</Danger>
           : null
         }
-      </Form>
+      </div>
     );
   }
 }
@@ -462,17 +161,13 @@ const mapDispatchToProps = (dispatch) => ({
   fetchRoles: () => { dispatch(fetchRoles(fetchRolesSuccess,fetchRolesFailure)) }
 });
 
+
 function mapStateToProps({ departments, users, roles }) {
   return {
     departments: departments.departments,
     users: users.allUsers,
-    roles: roles.roles
+    roles: roles.roles,
   };
 }
 
-UserForm = reduxForm({
-  form: 'user_form'
-})(UserForm);
-
-UserForm = withStyles(headerLinksStyle)(UserForm);
 export default UserForm = connect(mapStateToProps, mapDispatchToProps)(UserForm);
